Move client DeprecatedContractClass into a separate file
