+++ crates/starknet_client/src/objects/mod.rs
@@ -1,6 +1,7 @@
 pub mod block;
 #[cfg(test)]
 mod block_test;
+pub mod deprecated_contract_class;
 pub mod transaction;
 #[cfg(test)]
 mod transaction_test;
